app/src/test/java/com/dogbreedquiz/app/data/repository/DogBreedRepositoryImplTest.kt
package com.dogbreedquiz.app.data.repository

import com.dogbreedquiz.app.data.database.entity.BreedEntity
import com.dogbreedquiz.app.data.datasource.local.LocalBreedDataSource
import com.dogbreedquiz.app.data.datasource.remote.RemoteBreedDataSource
import com.dogbreedquiz.app.data.mapper.BreedMapper
import com.dogbreedquiz.app.domain.model.DogBreed
import io.mockk.coEvery
import io.mockk.coVerify
import io.mockk.mockk
import kotlinx.coroutines.flow.flowOf
import kotlinx.coroutines.test.runTest
import org.junit.Before
import org.junit.Test
import kotlin.test.assertEquals
import kotlin.test.assertNull
import kotlin.test.assertTrue

/**
 * Unit tests for DogBreedRepositoryImpl
 * Tests the repository implementation with proper mocking of dependencies
 */
class DogBreedRepositoryImplTest {

    private lateinit var remoteDataSource: RemoteBreedDataSource
    private lateinit var localDataSource: LocalBreedDataSource
    private lateinit var mapper: BreedMapper
    private lateinit var repository: DogBreedRepositoryImpl

    @Before
    fun setup() {
        remoteDataSource = mockk()
        localDataSource = mockk()
        mapper = mockk()
        repository = DogBreedRepositoryImpl(remoteDataSource, localDataSource, mapper)
    }

    @Test
    fun `getAllBreeds should return cached data when not force refresh and cache available`() = runTest {
        // Given
        val cachedEntities = listOf(createTestBreedEntity("1", "Golden Retriever"))
        val expectedBreeds = listOf(createTestBreed("1", "Golden Retriever"))

        coEvery { localDataSource.getAllBreeds() } returns cachedEntities
        coEvery { mapper.mapToDomain(cachedEntities[0]) } returns expectedBreeds[0]

        // When
        val result = repository.getAllBreeds(forceRefresh = false)

        // Then
        assertEquals(expectedBreeds, result)
        coVerify { localDataSource.getAllBreeds() }
        coVerify { mapper.mapToDomain(cachedEntities[0]) }
    }

    @Test
    fun `getAllBreeds should fetch from remote when force refresh is true`() = runTest {
        // Given
        val remoteEntities = listOf(createTestBreedEntity("1", "Golden Retriever"))
        val expectedBreeds = listOf(createTestBreed("1", "Golden Retriever"))

        coEvery { remoteDataSource.getAllBreeds() } returns remoteEntities
        coEvery { mapper.mapToDomain(remoteEntities[0]) } returns expectedBreeds[0]
        coEvery { localDataSource.saveBreeds(any()) } returns Unit

        // When
        val result = repository.getAllBreeds(forceRefresh = true)

        // Then
        assertEquals(expectedBreeds, result)
        coVerify { remoteDataSource.getAllBreeds() }
        coVerify { localDataSource.saveBreeds(remoteEntities) }
        coVerify { mapper.mapToDomain(remoteEntities[0]) }
    }

    @Test
    fun `getAllBreeds should return cached data when remote fails`() = runTest {
        // Given
        val cachedEntities = listOf(createTestBreedEntity("1", "Golden Retriever"))
        val expectedBreeds = listOf(createTestBreed("1", "Golden Retriever"))

        coEvery { localDataSource.getAllBreeds() } returns emptyList() andThen cachedEntities
        coEvery { remoteDataSource.getAllBreeds() } throws RuntimeException("Network error")
        coEvery { mapper.mapToDomain(cachedEntities[0]) } returns expectedBreeds[0]

        // When
        val result = repository.getAllBreeds(forceRefresh = false)

        // Then
        assertEquals(expectedBreeds, result)
        coVerify { mapper.mapToDomain(cachedEntities[0]) }
    }

    @Test
    fun `getAllBreeds should return fallback data when both remote and cache fail`() = runTest {
        // Given
        coEvery { localDataSource.getAllBreeds() } returns emptyList()
        coEvery { remoteDataSource.getAllBreeds() } throws RuntimeException("Network error")

        // When
        val result = repository.getAllBreeds(forceRefresh = false)

        // Then
        assertTrue(result.isNotEmpty()) // Should return fallback breeds
        assertTrue(result.any { it.name == "Golden Retriever" })
        assertTrue(result.any { it.name == "Labrador Retriever" })
    }

    @Test
    fun `getBreedById should return mapped breed when found`() = runTest {
        // Given
        val breedId = "golden_retriever"
        val breedEntity = createTestBreedEntity(breedId, "Golden Retriever")
        val expectedBreed = createTestBreed(breedId, "Golden Retriever")

        coEvery { localDataSource.getBreedById(breedId) } returns breedEntity
        coEvery { mapper.mapToDomain(breedEntity) } returns expectedBreed

        // When
        val result = repository.getBreedById(breedId)

        // Then
        assertEquals(expectedBreed, result)
        coVerify { localDataSource.getBreedById(breedId) }
        coVerify { mapper.mapToDomain(breedEntity) }
    }

    @Test
    fun `getBreedById should return null when breed not found`() = runTest {
        // Given
        val breedId = "nonexistent_breed"
        coEvery { localDataSource.getBreedById(breedId) } returns null

        // When
        val result = repository.getBreedById(breedId)

        // Then
        assertNull(result)
        coVerify { localDataSource.getBreedById(breedId) }
    }

    @Test
    fun `getAllBreedsFlow should return mapped flow from local data source`() = runTest {
        // Given
        val breedEntities = listOf(createTestBreedEntity("1", "Golden Retriever"))
        val expectedBreeds = listOf(createTestBreed("1", "Golden Retriever"))
        val entitiesFlow = flowOf(breedEntities)

        coEvery { localDataSource.getAllBreedsFlow() } returns entitiesFlow
        coEvery { mapper.mapToDomain(breedEntities[0]) } returns expectedBreeds[0]

        // When
        val resultFlow = repository.getAllBreedsFlow()

        // Then
        coVerify { localDataSource.getAllBreedsFlow() }
        // Note: Testing flows requires collecting values, which is more complex
        // In a real test, you'd collect the flow and verify the mapped values
    }

    @Test
    fun `loadBreedImage should return breed with image when loading succeeds`() = runTest {
        // Given
        val breedId = "golden_retriever"
        val breedEntity = createTestBreedEntity(breedId, "Golden Retriever", "")
        val updatedEntity = createTestBreedEntity(breedId, "Golden Retriever", "https://example.com/image.jpg")
        val expectedBreed = createTestBreed(breedId, "Golden Retriever")
        val imageUrl = "https://example.com/image.jpg"

        coEvery { localDataSource.getBreedById(breedId) } returns breedEntity andThen updatedEntity
        coEvery { remoteDataSource.getBreedImage(breedId) } returns imageUrl
        coEvery { mapper.mapToDomain(updatedEntity) } returns expectedBreed

        // When
        val result = repository.loadBreedImage(breedId)

        // Then
        assertEquals(expectedBreed, result)
        coVerify { localDataSource.updateBreedImage(breedId, imageUrl) }
        coVerify { remoteDataSource.getBreedImage(breedId) }
    }

    @Test
    fun `loadBreedImage should return breed without image when loading fails`() = runTest {
        // Given
        val breedId = "golden_retriever"
        val breedEntity = createTestBreedEntity(breedId, "Golden Retriever", "")
        val expectedBreed = createTestBreed(breedId, "Golden Retriever")

        coEvery { localDataSource.getBreedById(breedId) } returns breedEntity
        coEvery { remoteDataSource.getBreedImage(breedId) } throws RuntimeException("Network error")
        coEvery { mapper.mapToDomain(breedEntity) } returns expectedBreed

        // When
        val result = repository.loadBreedImage(breedId)

        // Then
        assertEquals(expectedBreed, result)
        coVerify { mapper.mapToDomain(breedEntity) }
    }

    @Test
    fun `searchBreeds should return mapped search results`() = runTest {
        // Given
        val query = "retriever"
        val searchEntities = listOf(createTestBreedEntity("1", "Golden Retriever"))
        val expectedBreeds = listOf(createTestBreed("1", "Golden Retriever"))

        coEvery { localDataSource.searchBreeds(query) } returns searchEntities
        coEvery { mapper.mapToDomain(searchEntities[0]) } returns expectedBreeds[0]

        // When
        val result = repository.searchBreeds(query)

        // Then
        assertEquals(expectedBreeds, result)
        coVerify { localDataSource.searchBreeds(query) }
        coVerify { mapper.mapToDomain(searchEntities[0]) }
    }

    private fun createTestBreedEntity(
        id: String,
        name: String,
        imageUrl: String = "https://example.com/image.jpg"
    ): BreedEntity {
        val currentTime = System.currentTimeMillis()
        return BreedEntity(
            id = id,
            name = name,
            imageUrl = imageUrl,
            description = "Test description",
            funFact = "Test fun fact",
            origin = "Test origin",
            size = "MEDIUM",
            temperament = "Friendly,Intelligent",
            lifeSpan = "10-12 years",
            difficulty = "BEGINNER",
            cachedAt = currentTime,
            expiresAt = currentTime + 7 * 24 * 60 * 60 * 1000L,
            lastUpdated = currentTime,
            apiSource = "dog.ceo",
            isFavorite = false
        )
    }

    private fun createTestBreed(
        id: String,
        name: String
    ): DogBreed {
        return DogBreed(
            id = id,
            name = name,
            imageUrl = "https://example.com/image.jpg",
            description = "Test description",
            funFact = "Test fun fact",
            origin = "Test origin",
            size = DogBreed.Size.MEDIUM,
            temperament = listOf("Friendly", "Intelligent"),
            lifeSpan = "10-12 years",
            difficulty = DogBreed.Difficulty.BEGINNER,
            isFavorite = false
        )
    }
}